requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
python-docx
requests
python-dotenv
filelock